refactor: split large files into focused modules
Break up runtime.ts, client-module-scope-analysis.ts,
client-module-transform.ts, vite-plugin/index.ts, adapter-node/index.ts,
cli/verification.ts, and cli/commands/generate.ts into smaller,
cohesive modules. Public APIs are unchanged.
