+++ .changeset/refactor-split-large-files.md
@@ -1,0 +1,8 @@
+---
+"@pracht/core": patch
+"@pracht/vite-plugin": patch
+"@pracht/adapter-node": patch
+"@pracht/cli": patch
+---
+
+Internal refactor: split several large modules into smaller, focused files to improve maintainability. Public APIs are unchanged.
