+++ packages/framework/src/prerender.ts
@@ -1,0 +1,132 @@
+import { buildPathFromSegments, resolveApp } from "./app.ts";
+import { resolveRegistryModule } from "./runtime-manifest.ts";
+import { handlePrachtRequest } from "./runtime.ts";
+import type {
+  ModuleRegistry,
+  PrachtApp,
+  ResolvedRoute,
+  RouteModule,
+  RouteRevalidate,
+} from "./types.ts";
+
+export interface PrerenderResult {
+  path: string;
+  html: string;
+  headers?: Record<string, string>;
+}
+
+export interface ISGManifestEntry {
+  revalidate: RouteRevalidate;
+}
+
+export interface PrerenderAppResult {
+  pages: PrerenderResult[];
+  isgManifest: Record<string, ISGManifestEntry>;
+}
+
+export interface PrerenderAppOptions {
+  app: PrachtApp;
+  registry?: ModuleRegistry;
+  clientEntryUrl?: string;
+  /** Per-source-file CSS map produced by the vite plugin. */
+  cssManifest?: Record<string, string[]>;
+  /** Per-source-file JS map produced by the vite plugin for modulepreload hints. */
+  jsManifest?: Record<string, string[]>;
+}
+
+export async function prerenderApp(options: PrerenderAppOptions): Promise<PrerenderResult[]>;
+export async function prerenderApp(
+  options: PrerenderAppOptions & { withISGManifest: true },
+): Promise<PrerenderAppResult>;
+export async function prerenderApp(
+  options: PrerenderAppOptions & { withISGManifest?: boolean },
+): Promise<PrerenderResult[] | PrerenderAppResult> {
+  const resolved = resolveApp(options.app);
+  const results: PrerenderResult[] = [];
+  const isgManifest: Record<string, ISGManifestEntry> = {};
+
+  // Collect all work items first, then render in parallel batches
+  const work: {
+    pathname: string;
+    render: string;
+    revalidate?: RouteRevalidate;
+  }[] = [];
+  for (const route of resolved.routes) {
+    if (route.render !== "ssg" && route.render !== "isg") continue;
+    const paths = await collectSSGPaths(route, options.registry);
+    for (const pathname of paths) {
+      work.push({ pathname, render: route.render, revalidate: route.revalidate });
+    }
+  }
+
+  const CONCURRENCY = 10;
+  for (let i = 0; i < work.length; i += CONCURRENCY) {
+    const batch = work.slice(i, i + CONCURRENCY);
+    const batchResults = await Promise.all(
+      batch.map(async (item) => {
+        const url = new URL(item.pathname, "http://localhost");
+        const request = new Request(url, { method: "GET" });
+
+        const response = await handlePrachtRequest({
+          app: options.app,
+          request,
+          registry: options.registry,
+          clientEntryUrl: options.clientEntryUrl,
+          cssManifest: options.cssManifest,
+          jsManifest: options.jsManifest,
+        });
+
+        if (response.status !== 200) {
+          console.warn(
+            `  Warning: ${item.render.toUpperCase()} route "${item.pathname}" returned status ${response.status}, skipping.`,
+          );
+          return null;
+        }
+
+        const html = await response.text();
+        return { headers: Object.fromEntries(response.headers), html, item };
+      }),
+    );
+
+    for (const result of batchResults) {
+      if (!result) continue;
+      results.push({
+        path: result.item.pathname,
+        html: result.html,
+        headers: result.headers,
+      });
+      if (result.item.render === "isg" && result.item.revalidate) {
+        isgManifest[result.item.pathname] = { revalidate: result.item.revalidate };
+      }
+    }
+  }
+
+  if (options.withISGManifest) {
+    return { pages: results, isgManifest };
+  }
+
+  return results;
+}
+
+async function collectSSGPaths(route: ResolvedRoute, registry?: ModuleRegistry): Promise<string[]> {
+  const hasDynamicSegments = route.segments.some(
+    (s) => s.type === "param" || s.type === "catchall",
+  );
+
+  if (!hasDynamicSegments) {
+    return [route.path];
+  }
+
+  // Dynamic route — must export getStaticPaths() to enumerate params
+  const routeModule = await resolveRegistryModule<RouteModule>(registry?.routeModules, route.file);
+
+  if (!routeModule?.getStaticPaths) {
+    console.warn(
+      `  Warning: SSG route "${route.path}" has dynamic segments but no getStaticPaths() export, skipping.`,
+    );
+    return [];
+  }
+
+  const paramSets = await routeModule.getStaticPaths();
+  return paramSets.map((params) => buildPathFromSegments(route.segments, params));
+}
